Fix issue with ClassList which was hitting an assert in debug mode https://bugs.webkit.org/show_bug.cgi?id=97820
Reviewed by Ojan Vafai.

http://trac.webkit.org/changeset/129779 hit asserts in debug mode when trying to use fastGetAttribute on an SVG element.

No new tests. No change in behavior.

* bindings/v8/V8Binding.h:
(WebCore::toNativeArguments):
* html/ClassList.h:

Canonical link: https://commits.webkit.org/115809@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@129798 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebCore/bindings/v8/V8Binding.h
@@ -222,13 +222,13 @@ namespace WebCore {
     }
 
     template <class T>
-    Vector<T> toNativeArguments(const v8::Arguments& args, size_t startIndex)
+    Vector<T> toNativeArguments(const v8::Arguments& args, int startIndex)
     {
         ASSERT(startIndex <= static_cast<size_t>(args.Length()));
         Vector<T> result;
         typedef NativeValueTraits<T> TraitsType;
-        size_t length = args.Length();
-        for (size_t i = startIndex; i < length; ++i)
+        int length = args.Length();
+        for (int i = startIndex; i < length; ++i)
             result.append(TraitsType::nativeValue(args[i]));
         return result;
     }
